# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EternaPy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List coding challenges
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
  /challenges/{challengeId}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get challenge detail
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
  /challenges/{challengeId}/submit:
    post:
      operationId: submitChallenge
      tags: [challenges]
      summary: Submit solution for a challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeSubmission"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /assignments/{assignmentId}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment detail
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentDetail"
  /assignments/{assignmentId}/submit:
    post:
      operationId: submitAssignment
      tags: [assignments]
      summary: Submit assignment for review
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeSubmission"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
...
  /ai/ask:
    post:
      operationId: askEternalAi
      tags: [ai]
      summary: Ask Eternal AI for hints and guidance (no direct answers)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiQuestion"
      responses:
        "200":
          description: AI response with hints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
...
components:
  schemas:
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
        input:
          type: string
        expectedOutput:
          type: string
        description:
          type: string
      required:
        - id
        - input
        - expectedOutput
        - description
...
    Challenge:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        language:
          type: string
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        tags:
          type: array
          items:
            type: string
        completionRate:
          type: number
      required:
        - id
        - title
        - description
        - difficulty
        - language
        - xpReward
        - isCompleted
        - tags
        - completionRate
    ChallengeDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        language:
          type: string
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        starterCode:
          type: string
        testCases:
          type: array
          items:
            $ref: "#/components/schemas/TestCase"
        hints:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - difficulty
        - language
        - xpReward
        - isCompleted
        - starterCode
        - testCases
        - hints
        - tags
    Assignment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        isPro:
          type: boolean
        courseId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - difficulty
        - xpReward
        - isCompleted
        - isPro
        - tags
    AssignmentDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        isPro:
          type: boolean
        starterCode:
          type: string
        language:
          type: string
        testCases:
          type: array
          items:
            $ref: "#/components/schemas/TestCase"
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - requirements
        - difficulty
        - xpReward
        - isCompleted
        - isPro
        - starterCode
        - language
        - testCases
        - tags
    CodeSubmission:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
      required:
        - code
        - language
    SubmissionResult:
      type: object
      properties:
        passed:
          type: boolean
        xpEarned:
          type: integer
        output:
          type: string
        error:
          type: ["string", "null"]
        testResults:
          type: array
          items:
            $ref: "#/components/schemas/TestResult"
        message:
          type: string
      required:
        - passed
        - xpEarned
        - output
        - testResults
        - message
    TestResult:
      type: object
      properties:
        passed:
          type: boolean
        description:
          type: string
        actualOutput:
          type: string
        expectedOutput:
          type: string
      required:
        - passed
        - description
        - actualOutput
        - expectedOutput
...
    AiQuestion:
      type: object
      properties:
        question:
          type: string
        context:
          type: string
        contextType:
          type: string
          enum: [lesson, challenge, assignment, general]
        code:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
      required:
        - question
        - contextType
    AiResponse:
      type: object
      properties:
        message:
          type: string
        hints:
          type: array
          items:
            type: string
        followUpQuestions:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
      required:
        - message
        - hints
        - followUpQuestions
        - resources
    Resource:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        type:
          type: string
      required:
        - title
        - url
        - type
...