# lib/api-zod/src/generated/types/aiResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EternaPy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from './resource';

export interface AiResponse {
  message: string;
  hints: string[];
  followUpQuestions: string[];
  resources: Resource[];
}