# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EternaPy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDifficulty } from './challengeDifficulty';

export interface Challenge {
  id: number;
  title: string;
  description: string;
  difficulty: ChallengeDifficulty;
  language: string;
  xpReward: number;
  isCompleted: boolean;
  tags: string[];
  completionRate: number;
}