# lib/api-zod/src/generated/types/challengeDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EternaPy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeDetailDifficulty } from './challengeDetailDifficulty';
import type { TestCase } from './testCase';

export interface ChallengeDetail {
  id: number;
  title: string;
  description: string;
  difficulty: ChallengeDetailDifficulty;
  language: string;
  xpReward: number;
  isCompleted: boolean;
  starterCode: string;
  testCases: TestCase[];
  hints: string[];
  tags: string[];
}