# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EternaPy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  title: string;
  url: string;
  type: string;
}