# lib/api-zod/src/generated/types/submissionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EternaPy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestResult } from './testResult';

export interface SubmissionResult {
  passed: boolean;
  xpEarned: number;
  output: string;
  /** @nullable */
  error?: string | null;
  testResults: TestResult[];
  message: string;
}